PizzaInfo entity fields fix

+++ src/main/java/groupId/artifactId/dao/PizzaInfoDao.java
@@ -164,7 +164,7 @@ public class PizzaInfoDao implements IPizzaInfoDao {
 
     private IPizzaInfo mapper(ResultSet resultSet) throws SQLException {
         return new PizzaInfo(resultSet.getLong("id"), resultSet.getString("name"),
-                resultSet.getString("description"), resultSet.getLong("size"),
+                resultSet.getString("description"), resultSet.getInt("size"),
                 resultSet.getTimestamp("creation_date").toLocalDateTime(), resultSet.getInt("version"));
     }
 }
